added tests to flask_tests
